+++ src/backend_qt/filter.rs
@@ -212,7 +212,10 @@ impl Filter<qt::Image> for QtFilter {
         ts: &usvg::Transform,
         input: Image,
     ) -> Result<Image, Error> {
-        let (dx, dy) = try_opt_or!(Self::resolve_offset(fe, units, bbox, ts), Ok(input));
+        let (dx, dy) = try_opt_or!(Self::scale_coordinates(fe.dx, fe.dy, units, bbox, ts), Ok(input));
+        if dx.is_fuzzy_zero() && dy.is_fuzzy_zero() {
+            return Ok(input);
+        }
 
         // TODO: do not use an additional buffer
         let mut buffer = create_image(input.width(), input.height())?;
@@ -502,7 +505,10 @@ impl Filter<qt::Image> for QtFilter {
         input: Image,
     ) -> Result<Image, Error> {
         let input = input.into_color_space(cs)?;
-        let (rx, ry) = try_opt_or!(Self::resolve_radius(fe, units, bbox, ts), Ok(input));
+        let (rx, ry) = try_opt_or!(
+            Self::scale_coordinates(fe.radius_x.value(), fe.radius_y.value(), units, bbox, ts),
+            Ok(input)
+        );
 
         let mut buffer = input.take()?;
 
@@ -512,6 +518,39 @@ impl Filter<qt::Image> for QtFilter {
                                   buffer.data_mut().as_bgra_mut());
 
         filter::from_premultiplied(buffer.data_mut().as_bgra_mut());
+
+        Ok(Image::from_image(buffer, cs))
+    }
+
+    fn apply_displacement_map(
+        fe: &usvg::FeDisplacementMap,
+        region: ScreenRect,
+        units: usvg::Units,
+        cs: ColorSpace,
+        bbox: Option<Rect>,
+        ts: &usvg::Transform,
+        input1: Image,
+        input2: Image,
+    ) -> Result<Image, Error> {
+        let input1 = input1.into_color_space(cs)?;
+        let input2 = input2.into_color_space(cs)?;
+        let (sx, sy) = try_opt_or!(
+            Self::scale_coordinates(fe.scale, fe.scale, units, bbox, ts),
+            Ok(input1)
+        );
+
+        let buffer1 = input1.take()?;
+        let buffer2 = input2.take()?;
+        let mut buffer = create_image(region.width(), region.height())?;
+
+        filter::displacement_map::apply(
+            fe.x_channel_selector, fe.y_channel_selector,
+            region.width(), region.height(),
+            sx, sy,
+            buffer1.data().as_bgra(),
+            buffer2.data().as_bgra(),
+            buffer.data_mut().as_bgra_mut(),
+        );
 
         Ok(Image::from_image(buffer, cs))
     }