+++ src/backend_skia/filter.rs
@@ -211,7 +211,10 @@ impl Filter<skia::Surface> for SkiaFilter {
         ts: &usvg::Transform,
         input: Image,
     ) -> Result<Image, Error> {
-        let (dx, dy) = try_opt_or!(Self::resolve_offset(fe, units, bbox, ts), Ok(input));
+        let (dx, dy) = try_opt_or!(Self::scale_coordinates(fe.dx, fe.dy, units, bbox, ts), Ok(input));
+        if dx.is_fuzzy_zero() && dy.is_fuzzy_zero() {
+            return Ok(input);
+        }
 
         let mut buffer = create_surface(input.width(), input.height())?;
 
@@ -538,7 +541,10 @@ impl Filter<skia::Surface> for SkiaFilter {
         use std::mem::swap;
 
         let input = input.into_color_space(cs)?;
-        let (rx, ry) = try_opt_or!(Self::resolve_radius(fe, units, bbox, ts), Ok(input));
+        let (rx, ry) = try_opt_or!(
+            Self::scale_coordinates(fe.radius_x.value(), fe.radius_y.value(), units, bbox, ts),
+            Ok(input)
+        );
 
         let mut buffer = input.take()?;
         let w = buffer.width();
@@ -560,6 +566,46 @@ impl Filter<skia::Surface> for SkiaFilter {
         if !skia::Surface::is_bgra() {
             data.as_bgra_mut().iter_mut().for_each(|p| swap(&mut p.r, &mut p.b));
         }
+
+        Ok(Image::from_image(buffer, cs))
+    }
+
+    fn apply_displacement_map(
+        fe: &usvg::FeDisplacementMap,
+        region: ScreenRect,
+        units: usvg::Units,
+        cs: ColorSpace,
+        bbox: Option<Rect>,
+        ts: &usvg::Transform,
+        input1: Image,
+        input2: Image,
+    ) -> Result<Image, Error> {
+        use std::mem::swap;
+
+        let input1 = input1.into_color_space(cs)?;
+        let input2 = input2.into_color_space(cs)?;
+        let (sx, sy) = try_opt_or!(
+            Self::scale_coordinates(fe.scale, fe.scale, units, bbox, ts),
+            Ok(input1)
+        );
+
+        let buffer1 = input1.take()?;
+        let mut buffer2 = input2.take()?;
+        let mut buffer = create_surface(region.width(), region.height())?;
+
+        // RGBA -> BGRA.
+        if !skia::Surface::is_bgra() {
+            buffer2.data_mut().as_bgra_mut().iter_mut().for_each(|p| swap(&mut p.r, &mut p.b));
+        }
+
+        filter::displacement_map::apply(
+            fe.x_channel_selector, fe.y_channel_selector,
+            region.width(), region.height(),
+            sx, sy,
+            buffer1.data().as_bgra(),
+            buffer2.data().as_bgra(),
+            buffer.data_mut().as_bgra_mut(),
+        );
 
         Ok(Image::from_image(buffer, cs))
     }