+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This changelog also contains important changes in dependencies.
 
 ## [Unreleased]
 ### Added
-- `feConvolveMatrix`, `feMorphology` support.
+- `feConvolveMatrix`, `feMorphology`, `feDisplacementMap` support.
 - Load grayscale raster images.
 - `enable-background` support.
 - `BackgroundImage`, `BackgroundAlpha`, `FillPaint` and `StrokePaint` support as a filter input.